Factor out spec-based construction path into factory area
