+++ python_modules/dagster/dagster/_core/definitions/decorators/asset_check_decorator.py
@@ -31,10 +31,10 @@ from dagster._core.execution.build_resources import wrap_resources_for_execution
 from dagster._utils.warnings import disable_dagster_warnings
 
 from ..input import In
-from .asset_decorator import (
+from .asset_decorator import make_asset_deps
+from .assets_definition_factory import (
     build_asset_ins,
     get_function_params_without_context_or_config_or_resources,
-    make_asset_deps,
 )
 from .op_decorator import _Op
