+++ python_modules/dagster/dagster/_core/definitions/decorators/asset_decorator.py
@@ -1,11 +1,8 @@
-from inspect import Parameter
 from typing import (
     AbstractSet,
     Any,
     Callable,
-    Dict,
     Iterable,
-    List,
     Mapping,
     Optional,
     Sequence,
@@ -18,9 +15,7 @@ from typing import (
 
 import dagster._check as check
 from dagster._annotations import deprecated_param, experimental_param
-from dagster._builtins import Nothing
-from dagster._config import UserConfigSchema
-from dagster._core.decorator_utils import get_function_params, get_valid_name_permutations
+from dagster._config.config_schema import UserConfigSchema
 from dagster._core.definitions.asset_dep import AssetDep, CoercibleToAssetDep
 from dagster._core.definitions.auto_materialize_policy import AutoMaterializePolicy
 from dagster._core.definitions.config import ConfigMapping
@@ -39,17 +34,18 @@ from ..asset_check_spec import AssetCheckSpec
 from ..asset_in import AssetIn
 from ..asset_out import AssetOut
 from ..asset_spec import AssetSpec
-from ..assets import ASSET_SUBSET_INPUT_PREFIX, AssetsDefinition, get_partition_mappings_from_deps
+from ..assets import AssetsDefinition, get_partition_mappings_from_deps
 from ..backfill_policy import BackfillPolicy, BackfillPolicyType
 from ..decorators.graph_decorator import graph
 from ..decorators.op_decorator import _Op
 from ..events import AssetKey, CoercibleToAssetKey, CoercibleToAssetKeyPrefix
-from ..input import GraphIn, In
+from ..input import GraphIn
 from ..output import GraphOut, Out
 from ..partition import PartitionsDefinition
 from ..policy import RetryPolicy
 from ..resource_definition import ResourceDefinition
 from ..utils import DEFAULT_IO_MANAGER_KEY, DEFAULT_OUTPUT, NoValueSentinel, validate_tags_strict
+from .assets_definition_factory import build_asset_ins, build_asset_outs
 
 
 @overload
@@ -652,20 +648,8 @@ def multi_asset(
 
         if asset_out_map and specs:
             raise DagsterInvalidDefinitionError("Must specify only outs or specs but not both.")
-        elif specs:
-            output_tuples_by_asset_key = {}
-            for asset_spec in specs:
-                output_name = asset_spec.key.to_python_identifier()
-                output_tuples_by_asset_key[asset_spec.key] = (
-                    output_name,
-                    Out(
-                        Nothing,
-                        is_required=not (can_subset or asset_spec.skippable),
-                        description=asset_spec.description,
-                        code_version=asset_spec.code_version,
-                        metadata=asset_spec.metadata,
-                    ),
-                )
+
+        if specs:
             if upstream_asset_deps:
                 raise DagsterInvalidDefinitionError(
                     "Can not pass deps and specs to @multi_asset, specify deps on the AssetSpecs"
@@ -677,77 +661,24 @@ def multi_asset(
                     " the AssetSpecs directly."
                 )
 
-            upstream_keys = set()
-            for spec in specs:
-                for dep in spec.deps:
-                    if dep.asset_key not in output_tuples_by_asset_key:
-                        upstream_keys.add(dep.asset_key)
-                    if (
-                        dep.asset_key in output_tuples_by_asset_key
-                        and dep.partition_mapping is not None
-                    ):
-                        # self-dependent asset also needs to be considered an upstream_key
-                        upstream_keys.add(dep.asset_key)
-
-            explicit_ins = ins or {}
-            # get which asset keys have inputs set
-            loaded_upstreams = build_asset_ins(fn, explicit_ins, deps=set())
-            unexpected_upstreams = {
-                key for key in loaded_upstreams.keys() if key not in upstream_keys
-            }
-            if unexpected_upstreams:
-                raise DagsterInvalidDefinitionError(
-                    f"Asset inputs {unexpected_upstreams} do not have dependencies on the passed"
-                    " AssetSpec(s). Set the deps on the appropriate AssetSpec(s)."
-                )
-            remaining_upstream_keys = {key for key in upstream_keys if key not in loaded_upstreams}
-            asset_ins = build_asset_ins(fn, explicit_ins, deps=remaining_upstream_keys)
-        else:
-            asset_ins = build_asset_ins(
-                fn,
-                ins or {},
-                deps=(
-                    {dep.asset_key for dep in upstream_asset_deps} if upstream_asset_deps else set()
-                ),
+            in_out_mapper = InOutMapper.from_specs(
+                specs=specs,
+                check_specs=check_specs or [],
+                can_subset=can_subset,
+                ins=ins or {},
+                fn=fn,
             )
-            output_tuples_by_asset_key = build_asset_outs(asset_out_map)
-
-            # validate that the asset_ins are a subset of the upstream asset_deps.
-            upstream_internal_asset_keys = set().union(*asset_deps.values())
-            asset_in_keys = set(asset_ins.keys())
-            if asset_deps and not asset_in_keys.issubset(upstream_internal_asset_keys):
-                invalid_asset_in_keys = asset_in_keys - upstream_internal_asset_keys
-                check.failed(
-                    f"Invalid asset dependencies: `{invalid_asset_in_keys}` specified as asset"
-                    " inputs, but are not specified in `internal_asset_deps`. Asset inputs must"
-                    " be associated with an output produced by the asset."
-                )
-
-            # validate that the asset_deps make sense
-            valid_asset_deps = asset_in_keys | set(output_tuples_by_asset_key.keys())
-            for out_name, asset_keys in asset_deps.items():
-                if asset_out_map and out_name not in asset_out_map:
-                    check.failed(
-                        f"Invalid out key '{out_name}' supplied to `internal_asset_deps` argument"
-                        f" for multi-asset {op_name}. Must be one of the outs for this multi-asset"
-                        f" {list(asset_out_map.keys())[:20]}.",
-                    )
-                invalid_asset_deps = asset_keys.difference(valid_asset_deps)
-                check.invariant(
-                    not invalid_asset_deps,
-                    f"Invalid asset dependencies: {invalid_asset_deps} specified in"
-                    f" `internal_asset_deps` argument for multi-asset '{op_name}' on key"
-                    f" '{out_name}'. Each specified asset key must be associated with an input to"
-                    " the asset or produced by this asset. Valid keys:"
-                    f" {list(valid_asset_deps)[:20]}",
-                )
-
-        in_out_mapper = InOutMapper.from_asset_ins_and_asset_outs(
-            asset_ins=asset_ins,
-            asset_outs=output_tuples_by_asset_key,
-            check_specs=check_specs or [],
-            can_subset=can_subset,
-        )
+        else:
+            in_out_mapper = InOutMapper.from_asset_outs(
+                asset_out_map=asset_out_map,
+                ins=ins or {},
+                fn=fn,
+                check_specs=check_specs or [],
+                op_name=op_name,
+                asset_deps=asset_deps,
+                upstream_asset_deps=upstream_asset_deps or [],
+                can_subset=can_subset,
+            )
 
         arg_resource_keys = {arg.name for arg in get_resource_args(fn)}
         check.param_invariant(
@@ -761,35 +692,13 @@ def multi_asset(
             f"Check output names overlap with asset output names: {in_out_mapper.overlapping_output_names}",
         )
 
-        if specs:
-            internal_deps = {
-                spec.key: {dep.asset_key for dep in spec.deps}
-                for spec in specs
-                if spec.deps is not None
-            }
-        else:
-            internal_deps = {
-                in_out_mapper.keys_by_output_name[name]: asset_deps[name] for name in asset_deps
-            }
-
-        # when a subsettable multi-asset is defined, it is possible that it will need to be
-        # broken into two separate parts, one which depends on the other. in order to represent
-        # this in the op execution graph, inputs need to be added to connect these possible deps
-        if can_subset and internal_deps:
-            asset_ins = {
-                **asset_ins,
-                **build_subsettable_asset_ins(
-                    asset_ins, output_tuples_by_asset_key, internal_deps.values()
-                ),
-            }
-
         with disable_dagster_warnings():
             op_required_resource_keys = required_resource_keys - arg_resource_keys
 
             op = _Op(
                 name=op_name,
                 description=description,
-                ins=dict(asset_ins.values()),
+                ins=in_out_mapper.asset_ins_by_input_names,
                 out=in_out_mapper.combined_outs_by_output_name,
                 required_resource_keys=op_required_resource_keys,
                 tags={
@@ -801,11 +710,8 @@ def multi_asset(
                 code_version=code_version,
             )(fn)
 
-        keys_by_input_name = {
-            input_name: asset_key for asset_key, (input_name, _) in asset_ins.items()
-        }
         partition_mappings = {
-            keys_by_input_name[input_name]: asset_in.partition_mapping
+            in_out_mapper.asset_keys_by_input_names[input_name]: asset_in.partition_mapping
             for input_name, asset_in in (ins or {}).items()
             if asset_in.partition_mapping is not None
         }
@@ -821,11 +727,11 @@ def multi_asset(
             resolved_specs = []
             input_deps_by_key = {
                 key: AssetDep(asset=key, partition_mapping=partition_mappings.get(key))
-                for key in keys_by_input_name.values()
+                for key in in_out_mapper.asset_keys_by_input_names.values()
             }
             input_deps = list(input_deps_by_key.values())
             for output_name, asset_out in asset_out_map.items():
-                key = in_out_mapper.keys_by_output_name[output_name]
+                key = in_out_mapper.asset_keys_by_output_name[output_name]
                 if internal_asset_deps:
                     deps = [
                         input_deps_by_key.get(
@@ -848,8 +754,8 @@ def multi_asset(
             resolved_specs = [spec._replace(group_name=group_name) for spec in resolved_specs]
 
         return AssetsDefinition.dagster_internal_init(
-            keys_by_input_name=keys_by_input_name,
-            keys_by_output_name=in_out_mapper.keys_by_output_name,
+            keys_by_input_name=in_out_mapper.asset_keys_by_input_names,
+            keys_by_output_name=in_out_mapper.asset_keys_by_output_name,
             node_def=op,
             partitions_def=partitions_def,
             can_subset=can_subset,
@@ -863,116 +769,6 @@ def multi_asset(
         )
 
     return inner
-
-
-def get_function_params_without_context_or_config_or_resources(
-    fn: Callable[..., Any],
-) -> List[Parameter]:
-    params = get_function_params(fn)
-    is_context_provided = len(params) > 0 and params[0].name in get_valid_name_permutations(
-        "context"
-    )
-    input_params = params[1:] if is_context_provided else params
-
-    resource_arg_names = {arg.name for arg in get_resource_args(fn)}
-
-    new_input_args = []
-    for input_arg in input_params:
-        if input_arg.name != "config" and input_arg.name not in resource_arg_names:
-            new_input_args.append(input_arg)
-
-    return new_input_args
-
-
-def stringify_asset_key_to_input_name(asset_key: AssetKey) -> str:
-    return "_".join(asset_key.path).replace("-", "_")
-
-
-def build_asset_ins(
-    fn: Callable[..., Any],
-    asset_ins: Mapping[str, AssetIn],
-    deps: Optional[AbstractSet[AssetKey]],
-) -> Mapping[AssetKey, Tuple[str, In]]:
-    """Creates a mapping from AssetKey to (name of input, In object)."""
-    deps = check.opt_set_param(deps, "deps", AssetKey)
-
-    new_input_args = get_function_params_without_context_or_config_or_resources(fn)
-
-    non_var_input_param_names = [
-        param.name for param in new_input_args if param.kind == Parameter.POSITIONAL_OR_KEYWORD
-    ]
-    has_kwargs = any(param.kind == Parameter.VAR_KEYWORD for param in new_input_args)
-
-    all_input_names = set(non_var_input_param_names) | asset_ins.keys()
-
-    if not has_kwargs:
-        for in_key, asset_in in asset_ins.items():
-            if in_key not in non_var_input_param_names and (
-                not isinstance(asset_in.dagster_type, DagsterType)
-                or not asset_in.dagster_type.is_nothing
-            ):
-                raise DagsterInvalidDefinitionError(
-                    f"Key '{in_key}' in provided ins dict does not correspond to any of the names "
-                    "of the arguments to the decorated function"
-                )
-
-    ins_by_asset_key: Dict[AssetKey, Tuple[str, In]] = {}
-    for input_name in all_input_names:
-        asset_key = None
-
-        if input_name in asset_ins:
-            asset_key = asset_ins[input_name].key
-            metadata = asset_ins[input_name].metadata or {}
-            key_prefix = asset_ins[input_name].key_prefix
-            input_manager_key = asset_ins[input_name].input_manager_key
-            dagster_type = asset_ins[input_name].dagster_type
-        else:
-            metadata = {}
-            key_prefix = None
-            input_manager_key = None
-            dagster_type = NoValueSentinel
-
-        asset_key = asset_key or AssetKey(list(filter(None, [*(key_prefix or []), input_name])))
-
-        ins_by_asset_key[asset_key] = (
-            input_name.replace("-", "_"),
-            In(metadata=metadata, input_manager_key=input_manager_key, dagster_type=dagster_type),
-        )
-
-    for asset_key in deps:
-        if asset_key in ins_by_asset_key:
-            raise DagsterInvalidDefinitionError(
-                f"deps value {asset_key} also declared as input/AssetIn"
-            )
-            # mypy doesn't realize that Nothing is a valid type here
-        ins_by_asset_key[asset_key] = (
-            stringify_asset_key_to_input_name(asset_key),
-            In(cast(type, Nothing)),
-        )
-
-    return ins_by_asset_key
-
-
-def build_subsettable_asset_ins(
-    asset_ins: Mapping[AssetKey, Tuple[str, In]],
-    asset_outs: Mapping[AssetKey, Tuple[str, Out]],
-    internal_upstream_deps: Iterable[AbstractSet[AssetKey]],
-) -> Mapping[AssetKey, Tuple[str, In]]:
-    """Creates a mapping from AssetKey to (name of input, In object) for any asset key that is not
-    currently an input, but may become one if this asset is subset.
-
-    For example, if a subsettable multi-asset produces both A and C, where C depends on both A and
-    some other asset B, there are some situations where executing just A and C without B will result
-    in these assets being generated by different steps within the same job. In this case, we need
-    a separate input to represent the fact that C depends on A.
-    """
-    # set of asset keys which are upstream of another asset, and are not currently inputs
-    potential_deps = set().union(*internal_upstream_deps).difference(set(asset_ins.keys()))
-    return {
-        key: (f"{ASSET_SUBSET_INPUT_PREFIX}{name}", In(Nothing))
-        for key, (name, _) in asset_outs.items()
-        if key in potential_deps
-    }
 
 
 @overload
@@ -1351,20 +1147,6 @@ def graph_multi_asset(
         )
 
     return inner
-
-
-def build_asset_outs(asset_outs: Mapping[str, AssetOut]) -> Mapping[AssetKey, Tuple[str, Out]]:
-    """Creates a mapping from AssetKey to (name of output, Out object)."""
-    outs_by_asset_key: Dict[AssetKey, Tuple[str, Out]] = {}
-    for output_name, asset_out in asset_outs.items():
-        out = asset_out.to_out()
-        asset_key = asset_out.key or AssetKey(
-            list(filter(None, [*(asset_out.key_prefix or []), output_name]))
-        )
-
-        outs_by_asset_key[asset_key] = (output_name.replace("-", "_"), out)
-
-    return outs_by_asset_key
 
 
 def _deps_and_non_argument_deps_to_asset_deps(
